+++ src/vue/pages/assets/AssetDetails.vue
@@ -210,6 +210,7 @@ export default {
       balances: vuexTypes.accountBalances,
       kvAssetTypeKycRequired: vuexTypes.kvAssetTypeKycRequired,
       isAccountUnverified: vuexTypes.isAccountUnverified,
+      isAccountCorporate: vuexTypes.isAccountCorporate,
     }),
     assetTermsUrl () {
       return this.asset.termsUrl(config.FILE_STORAGE)
@@ -218,8 +219,16 @@ export default {
       return !!this.balances.find(item => item.asset === this.asset.code)
     },
     isBalanceCreationAllowed () {
-      return this.asset.assetType === this.kvAssetTypeKycRequired &&
-        this.isAccountUnverified
+      switch (this.asset.assetType) {
+        case this.kycRequiredAssetType:
+          return !this.isAccountUnverified
+        case this.securityAssetType:
+          return this.isAccountGeneral ||
+                 this.isAccountUsAccredited ||
+                 this.isAccountCorporate
+        default:
+          return true
+      }
     },
   },
   async created () {
@@ -230,7 +239,7 @@ export default {
       loadBalances: vuexTypes.LOAD_ACCOUNT_BALANCES_DETAILS,
     }),
     async createBalance () {
-      if (this.isBalanceCreationAllowed) {
+      if (!this.isBalanceCreationAllowed) {
         Bus.error('asset-details.verification-required-err')
         return
       }