+++ src/vuex/types.js
@@ -10,6 +10,7 @@ const mutations = {
   SET_KV_ENTRY_BLOCKED_ROLE_ID: 'SET_KV_ENTRY_BLOCKED_ROLE_ID',
   SET_KV_ENTRY_US_VERIFIED_ROLE_ID: 'SET_KV_ENTRY_US_VERIFIED_ROLE_ID',
   SET_KV_ENTRY_US_ACCREDITED_ROLE_ID: 'SET_KV_ENTRY_US_ACCREDITED_ROLE_ID',
+  SET_KV_ASSET_TYPE_SECURITY: 'SET_KV_ASSET_TYPE_SECURITY',
 
   SET_KV_KYC_REQUIRED: 'SET_KV_KYC_REQUIRED',
   SET_DEFAULT_QUOTE_ASSET: 'SET_DEFAULT_QUOTE_ASSET',
@@ -37,6 +38,7 @@ const actions = {
   LOAD_KV_ENTRIES_ACCOUNT_ROLE_IDS: 'LOAD_KV_ENTRIES_ACCOUNT_ROLE_IDS',
   LOAD_KV_KYC_REQUIRED: 'LOAD_KV_KYC_REQUIRED',
   LOAD_DEFAULT_QUOTE_ASSET: 'LOAD_DEFAULT_QUOTE_ASSET',
+  LOAD_KV_ASSET_TYPE_SECURITY: 'LOAD_KV_ASSET_TYPE_SECURITY',
 
   // account
   LOAD_ACCOUNT: 'LOAD_ACCOUNT',
@@ -68,6 +70,7 @@ const getters = {
   kvEntryUsVerifiedRoleId: 'kvEntryUsVerifiedRoleId',
   kvEntryUsAccreditedRoleId: 'kvEntryUsAccreditedRoleId',
   kvAssetTypeKycRequired: 'kvAssetTypeKycRequired',
+  kvAssetTypeSecurity: 'kvAssetTypeSecurity',
   defaultQuoteAsset: 'defaultQuoteAsset',
 
   // account